requirements.txt
backports.ssl-match-hostname==3.5.0.1
certifi==2018.1.18
chardet==3.0.4
docker==3.1.0
docker-pycreds==0.2.2
future==0.16.0
idna==2.6
ipaddress==1.0.19
python-dateutil==2.6.1
pytimeparse==1.1.7
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
